pyproject.toml
[project]
name = "hackathone-v"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.83.0",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "markdown>=3.10.2",
    "openai-agents[litellm]>=0.9.2",
    "prometheus-client>=0.24.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "uvicorn>=0.41.0",
]
